reintegrate mGallery

+++ webEdition/liveUpdate/sqldumps/tblWorkflowDocTask.sql
@@ -1,9 +1,9 @@
 CREATE TABLE ###TBLPREFIX###tblWorkflowDocTask (
-  ID int(11) unsigned NOT NULL auto_increment,
-  documentStepID int(11) unsigned NOT NULL default '0',
-  workflowTaskID int(11) unsigned NOT NULL default '0',
-  `Date` int(10) unsigned NOT NULL default '0',
-  todoID int(11) unsigned NOT NULL default '0',
-  `Status` int(11) unsigned NOT NULL default '0',
+  ID int unsigned NOT NULL auto_increment,
+  documentStepID int unsigned NOT NULL default '0',
+  workflowTaskID int unsigned NOT NULL default '0',
+  `Date` int unsigned NOT NULL default '0',
+  todoID int unsigned NOT NULL default '0',
+  `Status` tinyint unsigned NOT NULL default '0',
   PRIMARY KEY  (ID)
 ) ENGINE=MyISAM;
